lesson1_intro/Config/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.2
djangorestframework==3.15.1
filelock==3.13.1
idna==3.6
MouseInfo==0.1.3
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyTelegramBotAPI==4.16.1
pytweening==1.0.7
requests==2.31.0
setuptools==69.1.1
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
